requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
Flask-CORS==5.0.0
pyngrok==7.2.0
Flask-SocketIO==5.3.4
mysql-connector-python==8.0.32
mariaDb==10.5.26
PyJWT==2.9.0
python_bcrypt==0.3.2
